+++ Tests/KeyboardKitTests/Actions/KeyboardAction+ReturnTests.swift
@@ -3,7 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2023-01-05.
-//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
 import KeyboardKit