+++ Tests/KeyboardKitTests/Emojis/Emoji+UnicodeTests.swift
@@ -3,7 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2022-12-12.
-//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
 //
 
 import KeyboardKit